+++ app/Http/Controllers/TimelinesController.php
@@ -5,17 +5,22 @@ use DB;
 use Auth;
 use Exception;
 use Throwable;
+
+use App\Http\Resources\PostCollection;
+use App\Http\Resources\Timeline as TimelineResource;
+use App\Models\User;
+use App\Libs\FeedMgr;
+use App\Libs\UserMgr;
+
+use App\Models\Setting;
+use App\Models\Timeline;
+use App\Models\Fanledger;
+use App\Models\Post;
+
+use Illuminate\Http\Request;
+use App\Enums\PaymentTypeEnum;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Http\Request;
-
-use App\Setting;
-use App\Libs\UserMgr;
-use App\Libs\FeedMgr;
-
-use App\Fanledger;
-use App\Timeline;
-use App\Enums\PaymentTypeEnum;
 
 class TimelinesController extends AppBaseController
 {
@@ -31,20 +36,56 @@ class TimelinesController extends AppBaseController
         ]);
     }
 
+    public function show(Request $request, Timeline $timeline)
+    {
+        $this->authorize('view', $timeline);
+        //$timeline->load(['avatar', 'cover']);
+        //$timeline->userstats = $request->user()->getStats();
+        //return [ 'timeline' => $timeline, ];
+        return new TimelineResource($timeline);
+    }
+
+    // Display my home timeline
+    public function homefeed(Request $request)
+    {
+        $query = Post::with('mediafiles', 'user')->withCount('comments')->where('active', 1);
+        $query->whereHas('timeline', function($q1) use(&$request) {
+            $q1->whereHas('followers', function($q2) use(&$request) {
+                $q2->where('users.id', $request->user()->id);
+            });
+        });
+        $data = $query->latest()->paginate( $request->input('take', env('MAX_POSTS_PER_REQUEST', 10)) );
+        return new PostCollection($data);
+    }
+
+    // Get a list of items that make up a timeline feed, typically posts but
+    //  keep generic as we may want to throw in other things
+    //  %TODO: 
+    //  ~ [ ] DEPRECATE, use FeedsController (?)
+    //  ~ [ ] trending tags
+    //  ~ [ ] announcements
+    //  ~ [ ] hashtag search
+    public function feed(Request $request, Timeline $timeline)
+    {
+        //$this->authorize('view', $timeline); // must be follower or subscriber
+        //$filters = [];
+        $query = Post::with('mediafiles', 'user')->withCount('comments')->where('active', 1);
+        $query->where('postable_type', 'timelines')->where('postable_id', $timeline->id);
+        $data = $query->latest()->paginate( $request->input('take', env('MAX_POSTS_PER_REQUEST', 10)) );
+        return new PostCollection($data);
+    }
+
     // Get suggested users (list/index)
     public function suggested(Request $request)
     {
         $TAKE = $request->input('take', 5);
-
-        $sessionUser = Auth::user();
-        $followedIDs = $sessionUser->followedtimelines->pluck('id');
-
-        $query = Timeline::with('user')->inRandomOrder();
-        $query->whereHas('user', function($q1) use(&$sessionUser, &$followedIDs) {
-            $q1->where('id', '<>', $sessionUser->id); // skip myself
+        $followedIDs = $request->user()->followedtimelines->pluck('id');
+        $query = Timeline::with(['user', 'avatar', 'cover'])->inRandomOrder();
+        $query->whereHas('user', function($q1) use(&$request, &$followedIDs) {
+            $q1->where('id', '<>', $request->user()->id); // skip myself
             // skip timelines I'm already following
-            $q1->whereHas('followedtimelines', function($q2) use(&$followedIDs) {
-                $q2->whereNotIn('timeline_id', $followedIDs);
+            $q1->whereDoesntHave('followedtimelines', function($q2) use(&$followedIDs) {
+                $q2->whereIn('shareable_id', $followedIDs);
             });
         });
 
@@ -56,87 +97,30 @@ class TimelinesController extends AppBaseController
         ]);
     }
 
-    public function show(Request $request, $username)
+    // %FIXME: better way to do this is to pull down a set of mediafiles associated with posts on this timeline (?)
+    public function previewPosts(Request $request, Timeline $timeline)
     {
-        $sessionUser = Auth::user();
-        $timeline = Timeline::with('user')->where('username', $username)->firstOrFail();
-        $sales = Fanledger::where('seller_id', $timeline->user->id)->sum('total_amount');
-
-        $timeline->userstats = [ // %FIXME DRY
-            'post_count' => $timeline->posts->count(),
-            'like_count' => 0, // %TODO $timeline->user->postlikes->count(),
-            'follower_count' => $timeline->followers->count(),
-            'following_count' => $timeline->user->followedtimelines->count(),
-            'subscribed_count' => 0, // %TODO $sessionUser->timeline->subscribed->count()
-            'earnings' => $sales,
-        ];
-
-        return view('timelines.show', [
-            'sessionUser' => $sessionUser,
-            'timeline' => $timeline,
-        ]);
+        $TAKE = $request->input('take', 6);
+        $query = Post::with('mediafiles', 'user')
+            ->has('mediafiles')
+            ->withCount('comments')->orderBy('comments_count', 'desc')
+            //->withCount('likes')->orderBy('likes_count', 'desc')
+            ->where('active', 1);
+        $query->where('postable_type', 'timelines')->where('postable_id', $timeline->id);
+        $data = $query->take($TAKE)->latest()->get();
+        return new PostCollection($data);
     }
 
-    // Display my home timeline
-    public function home(Request $request)
-    {
-        $sessionUser = Auth::user();
-        $timeline = $sessionUser->timeline()->with('user')->first();
-        $sales = Fanledger::where('seller_id', $timeline->user->id)->sum('total_amount');
 
-        $timeline->userstats = [ // %FIXME DRY
-            'post_count' => $timeline->posts->count(),
-            'like_count' => 0, // %TODO $timeline->user->postlikes->count(),
-            'follower_count' => $timeline->followers->count(),
-            'following_count' => $timeline->user->followedtimelines->count(),
-            'subscribed_count' => 0, // %TODO $sessionUser->timeline->subscribed->count()
-            'earnings' => $sales,
-        ];
-
-        return view('timelines.home', [
-            'sessionUser' => $sessionUser,
-            'timeline' => $timeline,
-            //'myVault' => $myVault,
-            //'vaultRootFolder' => $vaultRootFolder,
-        ]);
-    }
-
-    // Get a list of items that make up a timeline feed, typically posts but
-    //  keep generic as we may want to throw in other things
-    public function feeditems(Request $request, Timeline $timeline)
-    {
-        $sessionUser = Auth::user();
-        $follower = $timeline->user;
-        $page = $request->input('page', 1);
-        $take = 5; // $request->input('take', Setting::get('items_page'));
-
-        // %TODO
-        //  ~ [ ] trending tags
-        //  ~ [ ] announcements
-        //  ~ [ ] 
-
-        $filters = [];
-
-        if ($request->hashtag) {
-            $filters['hashtag'] = '#'.$request->hashtag;
-        }
-
-        $feeditems = FeedMgr::getPosts($follower, $filters, $page, $take); // %TODO: work with L5.8 pagination
-        //$feeditems = FeedMgr::getPostsRaw($sessionUser, $filters);
-
-        return response()->json([
-            'feeditems' => $feeditems,
-        ]);
-    }
-
+    // toggles, returns set state
     public function follow(Request $request, Timeline $timeline)
     {
-        $sessionUser = Auth::user(); // subscriber (purchaser)
+        $this->authorize('follow', $timeline);
 
         $request->validate([
-            'sharee_id' => 'required|numeric|min:1',
+            'sharee_id' => 'required|uuid|exists:users,id',
         ]);
-        if ( $request->sharee_id != $sessionUser->id ) {
+        if ( $request->sharee_id != $request->user()->id ) {
             abort(403);
         }
 
@@ -145,17 +129,25 @@ class TimelinesController extends AppBaseController
             $cattrs['notes'] = $request->notes;
         }
 
-        $timeline->followers()->detach($request->sharee_id); // remove existing
-        $timeline->followers()->attach($request->sharee_id, [ // will sync work here?
-            'shareable_type' => 'timelines',
-            'shareable_id' => $timeline->id,
-            'is_approved' => 1, // %FIXME
-            'access_level' => 'default',
-            'cattrs' => json_encode($cattrs),
-        ]); //
+        $existing = $timeline->followers->contains($request->sharee_id); // currently following?
+
+        if ($existing) {
+            $timeline->followers()->detach($request->sharee_id);
+            $isFollowing = false;
+        } else {
+            $timeline->followers()->attach($request->sharee_id, [ // will sync work here?
+                'shareable_type' => 'timelines',
+                'shareable_id' => $timeline->id,
+                'is_approved' => 1, // %FIXME
+                'access_level' => 'default',
+                'cattrs' => json_encode($cattrs),
+            ]); //
+            $isFollowing = true;
+        }
 
         $timeline->refresh();
         return response()->json([
+            'is_following' => $isFollowing,
             'timeline' => $timeline,
             'follower_count' => $timeline->followers->count(),
         ]);
@@ -163,48 +155,61 @@ class TimelinesController extends AppBaseController
 
     public function subscribe(Request $request, Timeline $timeline)
     {
-        $sessionUser = Auth::user(); // subscriber (purchaser)
+        $this->authorize('follow', $timeline);
 
         $request->validate([
-            'sharee_id' => 'required|numeric|min:1',
+            'sharee_id' => 'required|uuid|exists:users,id',
         ]);
-        if ( $request->sharee_id != $sessionUser->id ) {
+        if ( $request->sharee_id != $request->user()->id ) {
             abort(403);
         }
 
-        $timeline = DB::transaction( function() use(&$timeline, &$request, &$sessionUser) {
+        list($timeline, $isSubscribed) = DB::transaction( function() use(&$timeline, &$request) {
             $cattrs = [];
             if ( $request->has('notes') ) {
                 $cattrs['notes'] = $request->notes;
             }
-            $timeline->followers()->detach($request->sharee_id); // remove existing (covers 'upgrade') case
-            $timeline->followers()->attach($request->sharee_id, [
-                'shareable_type' => 'timelines',
-                'shareable_id' => $timeline->id,
-                'is_approved' => 1, // %FIXME
-                'access_level' => 'premium',
-                'cattrs' => json_encode($cattrs), // %FIXME: add a observer function?
-            ]); //
-            $timeline->receivePayment(
-                PaymentTypeEnum::SUBSCRIPTION,
-                $sessionUser,
-                $timeline->user->price*100, // %FIXME: should be on timeline
-                $cattrs,
-            );
-            return $timeline;
+
+            $existingFollowing = $timeline->followers->contains($request->sharee_id); // currently following?
+            $existingSubscribed = $timeline->subscribers->contains($request->sharee_id); // currently subscribed?
+
+            if ( $existingSubscribed ) {
+                // unsubscribe & unfollow
+                $timeline->followers()->detach($request->sharee_id);
+                $isSubscribed = false;
+            } else {
+                if ( $existingFollowing ) {
+                    // upgrade from follow to subscribe => remove existing (covers 'upgrade') case
+                    $timeline->followers()->detach($request->sharee_id); 
+                } // otherwise, just a direct subscription...
+                $timeline->followers()->attach($request->sharee_id, [
+                    'shareable_type' => 'timelines',
+                    'shareable_id' => $timeline->id,
+                    'is_approved' => 1, // %FIXME
+                    'access_level' => 'premium',
+                    'cattrs' => json_encode($cattrs), // %FIXME: add a observer function?
+                ]); //
+                $timeline->receivePayment(
+                    PaymentTypeEnum::SUBSCRIPTION,
+                    $request->user(),
+                    $timeline->price,
+                    $cattrs,
+                );
+                $isSubscribed = true;
+            }
+            return [$timeline, $isSubscribed];
         });
 
         $timeline->refresh();
         return response()->json([
+            'is_subscribed' => $isSubscribed,
             'timeline' => $timeline,
-            'follower_count' => $timeline->followers->count(),
+            'subscriber_count' => $timeline->subscribers->count(),
         ]);
     }
 
     public function tip(Request $request, Timeline $timeline)
     {
-        $sessionUser = Auth::user(); // sender of tip (purchaser)
-
         $request->validate([
             'base_unit_cost_in_cents' => 'required|numeric',
         ]);
@@ -217,7 +222,7 @@ class TimelinesController extends AppBaseController
         try {
             $timeline->receivePayment(
                 PaymentTypeEnum::TIP,
-                $sessionUser,
+                $request->user(),
                 $request->base_unit_cost_in_cents,
                 $cattrs,
             );
